Fix Xcode static analysis warnings
- Fix integer precision loss in isobmffbuffer.cpp by refactoring updateSampleDurationInternal() to take uint32_t parameter instead of uint64_t, matching ISOBMFF box constraints. Add validation at public API boundaries (setPtsAndDuration, truncate) to detect and log overflow conditions.

- Fix unused variable initialization in priv_aamp.cpp by removing redundant initialization of streamLockStopTime and licenseAcquisitionLockStopTime variables that are immediately overwritten before first use.

VPAAMP-888

+++ isobmff/isobmffbuffer.h
@@ -131,12 +131,12 @@ private:
 	 * @brief Private method to update the sample duration in the relevant boxes,
 	 *        if the sample duration is already present in those boxes.
 	 *
-	 * @param[in] duration - duration to set
+	 * @param[in] duration - duration to set (uint32_t per ISOBMFF spec)
 	 * @param[in] trun - Track fragment run box in which to update the duration, if present
 	 * @param[in] tfhd - Track fragment header box in which to update the duration, if present
 	 * @return true if the sample duration was updated in at least one box; false otherwise
 	 */
-	bool updateSampleDurationInternal(uint64_t duration, TrunBox& trun, TfhdBox& tfhd);
+	bool updateSampleDurationInternal(uint32_t duration, TrunBox& trun, TfhdBox& tfhd);
 
 	/**
 	 * @fn getBoxInfoInternal
